+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Bumps `JetBrains.Annotations` from 2023.2.0 to 2023.3.0
 - Bumps `Bogus` from 34.0.2 to 35.3.0
 - Bumps `Octokit` from 9.0.0 to 9.1.0
+- Bumps `FSharp.Core` from 8.0.100 to 8.0.101
 
 ## [1.6.0]
 ### Added